+++ R/report.R
@@ -15,7 +15,7 @@
 #' report(paper)
 #' }
 report <- function(paper,
-                   modules = c("imprecise_p", "marginal", "statcheck", "osf_check", "retractionwatch", "ref_consistency"),
+                   modules = c("exact_p", "marginal", "effect_size", "statcheck", "osf_check", "retractionwatch", "ref_consistency"),
                    output_file = paste0(paper$name, "_report.", output_format),
                    output_format = c("qmd", "html", "pdf")) {
   output_format <- match.arg(output_format)
@@ -147,7 +147,7 @@ report <- function(paper,
 #' @examples
 #' filename <- demoxml()
 #' paper <- read_grobid(filename)
-#' op <- module_run(paper, "imprecise_p")
+#' op <- module_run(paper, "exact_p")
 #' module_report(op) |> cat()
 module_report <- function(module_output,
                           header = 2,