+++ _stuff/validations/effect_size_ttest.R
@@ -16,28 +16,29 @@ detect_missing_effect_size_ttest <- function(paper, ...) {
   # Regex to detect all t-tests
   test_regex <- paste0(
     "\\bt\\s*", # word border and t
-    "(\\(\\s*\\d+(\\.\\d+)?\\s*\\))?", # df
+    "(\\(\\s*\\d+(\\.\\d+)?\\s*\\))?", # optional df
     "\\s*=\\s*", # comparator
     "[-+]?(\\d+(\\.\\d*)?|\\.\\d+)([eE][-+]?\\d+)?" # number
   )
   text_found_test <- paper |>
     search_text("=") |> # sentences with equal signs
     search_text("[0-9]") |> # sentences with numbers
-    search_text(test_regex, perl = TRUE) # sentences with a relevant test
+    # sentences with a relevant test
+    search_text(test_regex, perl = TRUE, ignore.case = FALSE)
+
 
   # Regex to detect effect sizes
   potentials <- c(
     "cohen('|\u2019)?s\\s+d",
     "d", "dz", "ds",
     "hedges?('|\u2019)?s?\\s+g",
-    "g",
-    "b",
-    "r",
-    "f\\s*(2|²)?",
-    "omega\\s*(2|²)?",
-    "ω\\s*(2|²)?",
-    "η\\s*p*\\s*(2|²)",
-    "partial\\s+η\\s*(2|²)"
+    "g", "b", "r"
+    # "cohen('|\u2019)?s\\s+f",
+    # "f\\s*(2|²)?",
+    # "omega\\s*(2|²)?",
+    # "ω\\s*(2|²)?",
+    # "η\\s*p*\\s*(2|²)",
+    # "partial\\s+η\\s*(2|²)"
   )
 
   es_regex <- paste0(